+++ src/onegov.user/onegov/user/__init__.py
@@ -10,19 +10,11 @@ from onegov.user.collections import UserCollection
 from onegov.user.collections import UserGroupCollection
 from onegov.user.models import User
 from onegov.user.models import UserGroup
-from onegov.user.utils import (
-    is_valid_yubikey,
-    is_valid_yubikey_format,
-    yubikey_otp_to_serial
-)
 
 __all__ = [
     'Auth',
-    'is_valid_yubikey',
-    'is_valid_yubikey_format',
     'User',
     'UserGroup',
     'UserCollection',
-    'UserGroupCollection',
-    'yubikey_otp_to_serial'
+    'UserGroupCollection'
 ]